+++ app/src/main/java/com/example/tapat/AdminActivity.java
@@ -9,6 +9,8 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 
+import com.example.tapat.adminfragments.AdminDashboard;
+
 public class AdminActivity extends AppCompatActivity {
 
     @Override
