+++ app/src/main/java/com/example/tapat/adminfragments/AdminCreate.java
@@ -1,4 +1,4 @@
-package com.example.tapat;
+package com.example.tapat.adminfragments;
 
 import android.os.Bundle;
 import android.text.InputType;
@@ -13,13 +13,8 @@ import android.widget.Spinner;
 import android.widget.TextView;
 
 import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentTransaction;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
 
-import java.util.ArrayList;
-import java.util.List;
+import com.example.tapat.R;
 
 public class AdminCreate extends Fragment {
