+++ app/src/main/java/com/example/tapat/adminfragments/AdminDashboard.java
@@ -1,4 +1,4 @@
-package com.example.tapat;
+package com.example.tapat.adminfragments;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -7,6 +7,8 @@ import android.view.ViewGroup;
 import android.widget.Button;
 
 import androidx.fragment.app.Fragment;
+
+import com.example.tapat.R;
 
 public class AdminDashboard extends Fragment {
