+++ app/src/main/java/com/example/tapat/adminfragments/AdminDisplayInfo.java
@@ -1,6 +1,5 @@
-package com.example.tapat;
+package com.example.tapat.adminfragments;
 
-import android.app.Activity;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -13,6 +12,8 @@ import android.widget.Spinner;
 import android.widget.TextView;
 
 import androidx.fragment.app.Fragment;
+
+import com.example.tapat.R;
 
 public class AdminDisplayInfo extends Fragment {
