+++ app/src/main/java/com/example/tapat/adminfragments/AdminList.java
@@ -1,4 +1,4 @@
-package com.example.tapat;
+package com.example.tapat.adminfragments;
 
 import android.os.Bundle;
 import android.text.Editable;
@@ -14,6 +14,9 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentTransaction;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.tapat.R;
+
 import java.util.Arrays;
 
 public class AdminList extends Fragment {
