+++ app/src/main/java/com/example/tapat/adminfragments/AdminNFCwriter.java
@@ -1,4 +1,4 @@
-package com.example.tapat;
+package com.example.tapat.adminfragments;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -8,24 +8,26 @@ import android.widget.TextView;
 
 import androidx.fragment.app.Fragment;
 
-public class AdminNFCReader extends Fragment {
+import com.example.tapat.R;
 
-    public AdminNFCReader() {
+public class AdminNFCwriter extends Fragment {
+
+    public AdminNFCwriter() {
         // Required empty public constructor
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.adminnfcreader, container, false);
+        View view = inflater.inflate(R.layout.adminnfcwriter, container, false);
 
         // Set the title in the TextView
-        TextView textView = view.findViewById(R.id.textViewNFCReader);
+        TextView textView = view.findViewById(R.id.textViewNFCWriter);
         if (textView != null) {
-            textView.setText("AdminNFCReader");
+            textView.setText("AdminNFCWriter");
         }
 
-        // Initialize and handle NFC Reader related UI components and functionality here
+        // Initialize and handle NFC Writer related UI components and functionality here
 
         return view;
     }
